requirements.txt
click>=8.1.3 
flet>=0.7.4
PyYAML>=6.0